pyproject.toml
[project]
name = "multi-ai-assistant-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.75.0",
    "fastapi[standard]>=0.124.2",
    "httpx>=0.28.1",
    "openai==0.28",
    "pydantic-settings>=2.12.0",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "streamlit>=1.52.1",
    "tiktoken>=0.12.0",
    "uvicorn[standard]>=0.38.0",
]
